PD-reglering fungerar som det ska

+++ styrenhet/Styrenhet/PDregulator.h
@@ -13,7 +13,7 @@ void PDactivate();
 void PDdeactivate();
 int PDisActive();
 void PDupdateSensorData(uint8_t left, uint8_t right);
-int8_t PDgetCorrection();
+int16_t PDgetCorrection();
 void setPD(uint8_t p_value, uint8_t d_value);
 void PDforward();
